fix: fix for wooahan's requirements
Delete interface
- Lotto
- LottoMachine

Rename Class
- KorLotto -> Lotto
- KorLottoMachine -> LottoMachine

Delete goal in README
- 입력과 출력에 대한 테스트 작성
