+++ README.md
@@ -3,7 +3,6 @@
 ## 목표
 - 객체지향적 설계
 - TDD와 같이 테스트를 먼저 만들고 테스트에 맞춰 구현
-- 입력과 출력에 대한 테스트 작성
 - git comment 일관성 있게 작성
 
 ## 요구 사항
