--- src/main/java/lotto/factory/LottoFactory.java
@@ -1,4 +1,0 @@
-package lotto.factory;
-
-public class LottoFactory {
-}
