--- src/main/java/lotto/model/KorLotto.java
@@ -1,20 +1,0 @@
-package lotto.model;
-
-import java.util.List;
-
-public class KorLotto {
-    private final List<Integer> numbers;
-
-    public KorLotto(List<Integer> numbers) {
-        validate(numbers);
-        this.numbers = numbers;
-    }
-
-    private void validate(List<Integer> numbers) {
-        if (numbers.size() != 6) {
-            throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
-        }
-    }
-
-    // TODO: 추가 기능 구현
-}
