--- src/main/java/lotto/model/KorLottoMachine.java
@@ -1,4 +1,0 @@
-package lotto.model;
-
-public class KorLottoMachine {
-}
