+++ src/main/java/lotto/model/Lotto.java
@@ -1,4 +1,20 @@
 package lotto.model;
 
-public interface Lotto {
+import java.util.List;
+
+public class Lotto {
+    private final List<Integer> numbers;
+
+    public Lotto(List<Integer> numbers) {
+        validate(numbers);
+        this.numbers = numbers;
+    }
+
+    private void validate(List<Integer> numbers) {
+        if (numbers.size() != 6) {
+            throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
+        }
+    }
+
+    // TODO: 추가 기능 구현
 }
