+++ src/main/java/lotto/model/LottoMachine.java
@@ -1,4 +1,4 @@
 package lotto.model;
 
-public interface LottoMachine {
+public class LottoMachine {
 }
